# app/src/main/webapp/app/shared/model/openapi/model/ApiOrchJobStats.ts
/**
 * API Measure Detector UI
 * Provides a bunch of useful REST Endpoints for the Measure Detector UI
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ApiOrchJobStats {
  id?: number;
  total?: number;
  scheduled?: number;
  enqueued?: number;
  processing?: number;
  failed?: number;
  succeeded?: number;
}